+++ mindsdb/integrations/handlers/utilities/auth_utilities/google/__init__.py
@@ -1,2 +1,2 @@
-from .google_oauth_utilities import GoogleUserOAuth2Manager
+from .google_user_oauth_utilities import GoogleUserOAuth2Manager
 from .google_service_account_oauth_utilities import GoogleServiceAccountOAuth2Manager
